Remove changing gTTA depth logic, default to 0 depth
Depth was not used in bootstrap phase, as the genesis transaction is always
explicitly attached to, even if depth was adjusted at every cycle to
minDepth.
In addition, all the updateDepth logic is unreliable and might have
unexpected results: removed min and max depth ranges configuration
options and just rely on one and fixed initial depth value.